Added a class to manage the game pasued text
Added a new class to manage the game paused text that appears in the pause screen.
